refactor(lab): decompose secrets god-file test module into submodules
Re-applied at the crates/homeboy-core path after #8389 extracted the
homeboy-core crate.

runner/lab/secrets.rs was 2354 lines (god_file) with ~962 lines of
production and a ~1390-line inline test module. Split the test module
into secrets/tests/{part_a,part_b}.rs with shared helpers (run_plan_args,
secret_env_plan_from_args, fixture_runner, RemovedEnvVar, fixture
providers) promoted to secrets/tests/mod.rs as pub(super).

secrets.rs is now 964 lines (production only). cargo check -p
homeboy-core --lib is clean. Clears the god_file finding.
